Rename project to Fast Quartz
