+++ fast-quartz-core/build.gradle.kts
@@ -1,0 +1,20 @@
+plugins {
+    `java-library`
+}
+
+base {
+    archivesName.set("fast-quartz-core")
+}
+
+dependencies {
+    api(project(":fabric-stubs"))
+    implementation("org.slf4j:slf4j-api:2.0.13")
+    testRuntimeOnly("org.slf4j:slf4j-simple:2.0.13")
+}
+
+tasks.processResources {
+    inputs.property("version", project.version)
+    filesMatching("fabric.mod.json") {
+        expand(mapOf("version" to project.version))
+    }
+}
