+++ fast-quartz-core/src/main/resources/fabric.mod.json
@@ -1,0 +1,22 @@
+{
+  "schemaVersion": 1,
+  "id": "fastquartz",
+  "version": "${version}",
+  "name": "Fast Quartz Core",
+  "description": "Redstone-accurate simulation engine foundation.",
+  "authors": ["Fast Quartz Team"],
+  "contact": {
+    "homepage": "https://github.com/fast-quartz/fast-quartz",
+    "sources": "https://github.com/fast-quartz/fast-quartz"
+  },
+  "license": "MIT",
+  "environment": "*",
+  "entrypoints": {
+    "main": ["dev.fastquartz.fastquartz.FastQuartzMod"]
+  },
+  "depends": {
+    "fabricloader": ">=0.15.11",
+    "minecraft": "1.20.1",
+    "java": ">=17"
+  }
+}
