+++ fast-quartz-core/src/test/java/dev/fastquartz/fastquartz/FastQuartzModTest.java
@@ -1,0 +1,12 @@
+package dev.fastquartz.fastquartz;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
+import org.junit.jupiter.api.Test;
+
+class FastQuartzModTest {
+  @Test
+  void modIdIsStable() {
+    assertEquals("fastquartz", FastQuartzMod.MOD_ID);
+  }
+}
