--- faststone-core/src/main/resources/fabric.mod.json
@@ -1,22 +1,0 @@
-{
-  "schemaVersion": 1,
-  "id": "faststone",
-  "version": "${version}",
-  "name": "Faststone Core",
-  "description": "Redstone-accurate simulation engine foundation.",
-  "authors": ["Faststone Team"],
-  "contact": {
-    "homepage": "https://github.com/faststone-dev/fast-quartz",
-    "sources": "https://github.com/faststone-dev/fast-quartz"
-  },
-  "license": "MIT",
-  "environment": "*",
-  "entrypoints": {
-    "main": ["dev.fastquartz.faststone.FaststoneMod"]
-  },
-  "depends": {
-    "fabricloader": ">=0.15.11",
-    "minecraft": "1.20.1",
-    "java": ">=17"
-  }
-}
