--- faststone-core/src/test/java/dev/fastquartz/faststone/FaststoneModTest.java
@@ -1,12 +1,0 @@
-package dev.fastquartz.faststone;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-
-import org.junit.jupiter.api.Test;
-
-class FaststoneModTest {
-  @Test
-  void modIdIsStable() {
-    assertEquals("faststone", FaststoneMod.MOD_ID);
-  }
-}
