+++ headless-runner/build.gradle.kts
@@ -8,7 +8,7 @@ base {
 
 dependencies {
     implementation(project(":fabric-stubs"))
-    implementation(project(":faststone-core"))
+    implementation(project(":fast-quartz-core"))
     implementation("org.slf4j:slf4j-api:2.0.13")
     runtimeOnly("org.slf4j:slf4j-simple:2.0.13")
 }
