+++ headless-runner/src/main/resources/fabric.mod.json
@@ -1,10 +1,10 @@
 {
   "schemaVersion": 1,
-  "id": "faststone-headless",
+  "id": "fastquartz-headless",
   "version": "${version}",
-  "name": "Faststone Headless Runner",
-  "description": "Headless dedicated server wrapper for Faststone.",
-  "authors": ["Faststone Team"],
+  "name": "Fast Quartz Headless Runner",
+  "description": "Headless dedicated server wrapper for Fast Quartz.",
+  "authors": ["Fast Quartz Team"],
   "license": "MIT",
   "environment": "server",
   "entrypoints": {
@@ -14,6 +14,6 @@
     "fabricloader": ">=0.15.11",
     "minecraft": "1.20.1",
     "java": ">=17",
-    "faststone": "*"
+    "fastquartz": "*"
   }
 }
